Redid the color trans function to solve some issues, and incorporate some of the aspects of Aarons function.
This function adopts the path calculation approach, changes the calculation to use divmod instead. This prevents rounding errors, but depending on the number of steps the last step can be a significant change.

This would be less of an issue if temperature is set in K because (fabs(path) % steps should be a comparitively small jump.

This function also uses the command method for setting the values, though this seems to break if you attempt to send commands too quickly.